Moves logic and functions of location update to the new child component. Child component renders succesfully.
